Preference min max fields
Adding minPreference and maxPreference to handle scaling of negative and positive preference range data.
These values will be calculated per preference data setup. They represent the range of negative and postitive preference data.
When rendering the heatmaps we will divide the negative values by abs(minPreference) and positive by maxPreference.
This should give is a number between 0 and 1 that we can use to determine the intensity of the rect in the heatmap.