underline core in prediction details
Adding core_offset and core_length to model preferenceconf.yaml.
Optional tags that allow user to specify where in a prediction the core will be.
There will odd sized predictions in the future so this was not something we can calculate.
Drawing the underline using border so it will be a consistent color(black).