smaller db through compact indexes
Adds new gene_id field to gene table.
gene_id is built from chrom, gene_begin, common_name, strand and ordered that way.
Within the gene lists we download there is no unique identifier for what we consider to be a gene.
There is a gene symbol and id but we consider a gene to be a gene_symbol + TSS.

Loading code will now populate this new field.
Queries will make use of this field.

Also fixes an issue with sort by max + preference data and paging related to the same gene symbol on multiple chroms.